+++ web/modules/custom/gos_migrate/src/Plugin/migrate/process/ContextualLinksMapper.php
@@ -62,7 +62,7 @@ class ContextualLinksMapper extends ProcessPluginBase {
 
       $value = $row->getSourceProperty($field);
 
-      if (empty($value)) {
+      if ($value === NULL || $value === '') {
         continue;
       }