+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update drupal/core-dev (10.2.3 => 10.3.3)
 - update all minors (25) dependencies
 - update drush/drush (12.5.3 => 13.1.1)
+- update vimeo/psalm (4.30.0 => 5.26.1)
 
 ### Changed
 - replace all deprecated usage of MigrateSkipProcessException by stopPipeline & ->saveMessage